Fix HHH-10550
Fixes 
https://hibernate.atlassian.net/browse/HHH-10550

+++ hibernate-core/src/main/java/org/hibernate/engine/jdbc/internal/BinaryStreamImpl.java
@@ -18,7 +18,7 @@ import org.hibernate.engine.jdbc.BinaryStream;
  * @author Steve Ebersole
  */
 public final class BinaryStreamImpl extends ByteArrayInputStream implements BinaryStream {
-	private final int length;
+	private final long length;
 
 	/**
 	 * Constructs a BinaryStreamImpl
